Add files via upload
This is a demo Hospital website which has been made for the Dashboard Analytics for Hospital Management with the help of IBM tool called as Cognos BI. the link of the Dashboard has been attached in the website's Dashboard button. If you wanna see my analytics in Hospital Management (Heart Disease) go and check it out!